float

+++ src/parser/DependencyParser.java
@@ -32,10 +32,10 @@ public class DependencyParser implements Serializable {
 	
 	DependencyParser pruner;
 	
-	double pruningGoldHits = 0;
-	double pruningTotGold = 1e-30;
-	double pruningTotUparcs = 0;
-	double pruningTotArcs = 1e-30;
+	float pruningGoldHits = 0;
+	float pruningTotGold = 1e-30f;
+	float pruningTotUparcs = 0;
+	float pruningTotArcs = 1e-30f;
 	
 	public static void main(String[] args) 
 		throws IOException, ClassNotFoundException, CloneNotSupportedException
@@ -54,8 +54,8 @@ public class DependencyParser implements Serializable {
 			prunerOptions.pruning = false;
 			prunerOptions.test = false;
 			prunerOptions.learnLabel = false;
-			prunerOptions.gamma = 1.0;
-			prunerOptions.gammaLabel = 1.0;
+			prunerOptions.gamma = 1.0f;
+			prunerOptions.gammaLabel = 1.0f;
 			prunerOptions.R = 0;
 			prunerOptions.bits = options.bits - 2;
 			
@@ -120,13 +120,13 @@ public class DependencyParser implements Serializable {
 	{
 		if (options.numTestConverge < 10) return;
 		System.out.println(" Tuning hill-climbing converge number on eval set...");
-		double maxUAS = evaluateWithConvergeNum(options.numTrainConverge);
+		float maxUAS = evaluateWithConvergeNum(options.numTrainConverge);
 		System.out.printf("\tconverge=%d\tUAS=%f%n", options.numTrainConverge, maxUAS);
 		int max = options.numTrainConverge / 5;
 		int min = 2;
 		while (min < max) {
 			int mid = (min+max)/2;
-			double uas = evaluateWithConvergeNum(mid*5);
+			float uas = evaluateWithConvergeNum(mid*5);
 			System.out.printf("\tconverge=%d\tUAS=%f%n", mid*5, uas);
 			if (uas + 0.001 < maxUAS)
 				min = mid+1;
@@ -173,7 +173,7 @@ public class DependencyParser implements Serializable {
 				pruningTotUparcs / pruningTotArcs);
 	}
 	
-	public double pruningRecall()
+	public float pruningRecall()
 	{
 		return pruningGoldHits / pruningTotGold;
 	}
@@ -181,9 +181,9 @@ public class DependencyParser implements Serializable {
 	public void resetPruningStats()
 	{
 		pruningGoldHits = 0;
-		pruningTotGold = 1e-30;
+		pruningTotGold = 1e-30f;
 		pruningTotUparcs = 0;
-		pruningTotArcs = 1e-30;
+		pruningTotArcs = 1e-30f;
 	}
 	
     public void train(DependencyInstance[] lstTrain) 
@@ -196,12 +196,12 @@ public class DependencyParser implements Serializable {
         	Options optionsBak = (Options) options.clone();
         	options.learningMode = LearningMode.Basic;
         	options.R = 0;
-        	options.gamma = 1.0;
-        	options.gammaLabel = 1.0;
+        	options.gamma = 1.0f;
+        	options.gammaLabel = 1.0f;
         	options.maxNumIters = options.numPretrainIters;
             options.useHO = false;
-        	parameters.gamma = 1.0;
-        	parameters.gammaLabel = 1.0;
+        	parameters.gamma = 1.0f;
+        	parameters.gammaLabel = 1.0f;
         	parameters.rank = 0;
     		System.out.println("=============================================");
     		System.out.printf(" Pre-training:%n");
@@ -278,7 +278,7 @@ public class DependencyParser implements Serializable {
             int offset = (N % 3 == 0) ? iIter : 0;
 
     		long start = 0;
-    		double loss = 0;
+    		float loss = 0;
     		int uas = 0, tot = 0;
     		start = System.currentTimeMillis();
                 		    		
@@ -336,7 +336,7 @@ public class DependencyParser implements Serializable {
                 	parameters.averageParameters((iIter+1)*N);
                 int cnvg = options.numTestConverge;
                 options.numTestConverge = options.numTrainConverge;
-	  			double res = evaluateSet(false, false);
+	  			float res = evaluateSet(false, false);
                 options.numTestConverge = cnvg;
                 System.out.println();
 	  			System.out.println("_____________________________________________");
@@ -401,7 +401,7 @@ public class DependencyParser implements Serializable {
     	return nCorrect;
     }
     
-    public double evaluateSet(boolean output, boolean evalWithPunc)
+    public float evaluateSet(boolean output, boolean evalWithPunc)
     		throws IOException {
     	
     	if (pruner != null) pruner.resetPruningStats();
@@ -465,7 +465,7 @@ public class DependencyParser implements Serializable {
         return eval.UAS();
     }
     
-    public double evaluateWithConvergeNum(int converge) throws IOException, CloneNotSupportedException 
+    public float evaluateWithConvergeNum(int converge) throws IOException, CloneNotSupportedException 
     {
     	
     	if (pruner != null) pruner.resetPruningStats();
